requirements.txt
httpx>=0.27.0
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.2.0
pandas>=2.2.0
python-dotenv>=1.0.1
playwright>=1.44.0
